refactor: 공통 hook들을 hooks/common으로 이동
