+++ frontend/src/components/Review/ReviewImageUploader/ReviewImageUploader.tsx
@@ -2,7 +2,7 @@ import { Button, Heading, Spacing, Text, useTheme } from '@fun-eat/design-system
 import type { ChangeEventHandler } from 'react';
 import styled from 'styled-components';
 
-import useEnterKeyDown from '@/hooks/useEnterKeyDown';
+import { useEnterKeyDown } from '@/hooks/common';
 
 interface ReviewImageUploaderProps {
   reviewPreviewImage: string;
