+++ frontend/src/hooks/common/useIntersectionObserver.ts
